WIP: more refactoring
